Speed up `ein.array`, add `use_cache=False`
`ein.array` now accepts `use_cache=False` to avoid calling the slow
`lambda_hash` in cases where we might want to avoid it.

Additionally, instead of eagerly applying the function twice to get the
target device, we speed things up by up to a factor of 2 in the common
case where either the default device is correct or `device=` is
explicitly passed.  We catch `RuntimeError` and parse the message for
`same device` to know when to try correcting for failure, and emit a
warning in such cases.